+++ examples/simulation-study/src/llhd-profiles.R
@@ -21,6 +21,8 @@ INPUT_FILE <- config$outputLlhdFile
 
 if (!file.exists(INPUT_FILE)) {
     stop(sprintf("Cannot find the input file: %s", INPUT_FILE))
+} else {
+    cat("Found the input file:", INPUT_FILE, "\n")
 }
 
 x <- read.csv(INPUT_FILE, header = FALSE, stringsAsFactors = FALSE)
@@ -40,13 +42,14 @@ fig_theme <- theme(
     legend.title = element_text(size = font_scale_factor * 22),
     plot.title = element_text(size = font_scale_factor * 32),
     plot.subtitle = element_text(size = font_scale_factor * 22),
-    panel.grid.minor = ggplot2::element_blank(),
+    panel.grid.minor.x = ggplot2::element_blank(),
+    panel.grid.minor.y = ggplot2::element_blank(),
+    ## panel.grid.minor.y = ggplot2::element_line(color="#cbcbcb"),
     panel.grid.major.y = ggplot2::element_line(color="#cbcbcb"),
     panel.grid.major.x = ggplot2::element_blank(),
     panel.background = ggplot2::element_blank(),
     strip.background = ggplot2::element_rect(fill="white"))
 
-y_scaling <- scale_y_continuous(breaks = seq(from = -60, to = -45, by = 5))
 
 truth_linetype <- "dashed"
 
@@ -55,7 +58,7 @@ lambda_figure <- ggplot(filter(x, mu == PARAMS$mu, psi == PARAMS$psi, rho == PAR
     geom_vline(xintercept = PARAMS$lambda, linetype = truth_linetype) +
     labs(x = "Birth rate",
          y = "Log-Likelihood") +
-    y_scaling +
+    scale_y_continuous(breaks = seq(from = -1e3, to = 1e3, by = 5)) +
     fig_theme
 
 if (save_figures) {
@@ -69,7 +72,7 @@ mu_figure <- ggplot(filter(x, lambda == PARAMS$lambda, psi == PARAMS$psi, rho ==
     geom_vline(xintercept = PARAMS$mu, linetype = truth_linetype) +
     labs(x = "Death rate",
          y = "Log-Likelihood") +
-    y_scaling +
+    scale_y_continuous(breaks = seq(from = -1e3, to = 1e3, by = 5)) +
     fig_theme
 
 if (save_figures) {
@@ -83,7 +86,7 @@ psi_figure <- ggplot(filter(x, lambda == PARAMS$lambda, mu == PARAMS$mu, rho == 
     geom_vline(xintercept = PARAMS$psi, linetype = truth_linetype) +
     labs(x = "Sampling rate",
          y = "Log-Likelihood") +
-    y_scaling +
+    scale_y_continuous(breaks = seq(from = -1e3, to = 1e3, by = 5)) +
     fig_theme
 
 if (save_figures) {
@@ -97,7 +100,7 @@ rho_figure <- ggplot(filter(x, lambda == PARAMS$lambda, mu == PARAMS$mu, psi == 
     geom_vline(xintercept = PARAMS$rho, linetype = truth_linetype) +
     labs(x = "Catastrophe extinction\nprobability",
          y = "Log-Likelihood") +
-    y_scaling +
+    scale_y_continuous(breaks = seq(from = -1e3, to = 1e3, by = 5)) +
     fig_theme
 
 if (save_figures) {
@@ -111,7 +114,7 @@ omega_figure <- ggplot(filter(x, lambda == PARAMS$lambda, mu == PARAMS$mu, psi =
     geom_vline(xintercept = PARAMS$omega, linetype = truth_linetype) +
     labs(x = "Occurrence rate",
          y = "Log-Likelihood") +
-    y_scaling +
+    scale_y_continuous(breaks = seq(from = -1e3, to = 1e3, by = 5)) +
     fig_theme
 
 if (save_figures) {
@@ -125,7 +128,7 @@ nu_figure <- ggplot(filter(x, lambda == PARAMS$lambda, mu == PARAMS$mu, psi == P
     geom_vline(xintercept = PARAMS$nu, linetype = truth_linetype) +
     labs(x = "Disaster extinction\nprobability",
          y = "Log-Likelihood") +
-    y_scaling +
+    scale_y_continuous(breaks = seq(from = -1e3, to = 1e3, by = 5)) +
     fig_theme
 
 if (save_figures) {
@@ -140,16 +143,17 @@ if (save_figures) {
 simulationEvents <- suppressWarnings(readLines(config$outputEventsFile))
 et <- as.list(table(str_extract(simulationEvents, "^[a-zA-Z]+")))
 
-mask <- grepl(pattern = "CatastropheEvent", x = simulationEvents)
+mask <- grepl(pattern = "Catastrophe", x = simulationEvents)
 num_catastropheed <- sapply(str_match_all(simulationEvents[mask], pattern = "Person\\s[0-9]+"), length)
 rm(mask)
 
-mask <- grepl(pattern = "DisasterEvent", x = simulationEvents)
+mask <- grepl(pattern = "Disaster", x = simulationEvents)
 num_disastered <- sapply(str_match_all(simulationEvents[mask], pattern = "Person\\s[0-9]+"), length)
 rm(mask)
 
+num_unobserved_lineages <- 1 + et$Infection - et$Occurrence - et$Removal - et$Sampling - sum(num_catastropheed) - sum(num_disastered)
+cat(sprintf("\nThe number of unobserved lineages, a.k.a. the final prevalence, is %d\n\n", num_unobserved_lineages))
 
-num_unobserved_lineages <- 1 + et$InfectionEvent - et$OccurrenceEvent - et$RemovalEvent - et$SamplingEvent - sum(num_catastropheed) - sum(num_disastered)
 
 ## Because we do not include the true parameters in the LLHD evaluations, we
 ## need to get the closest match to them among those that we have.
@@ -157,7 +161,7 @@ unique_thresh <- 0.0006
 curr_nb <- as.list(filter(x, abs(lambda - PARAMS$lambda) < unique_thresh, mu == PARAMS$mu, psi == PARAMS$psi, abs(rho - PARAMS$rho) < unique_thresh, omega == PARAMS$omega, abs(nu - PARAMS$nu) < unique_thresh) %>% select(starts_with("neg_binom")))
 curr_nb <- list(neg_binom_r = mean(curr_nb$neg_binom_r), neg_binom_p = mean(curr_nb$neg_binom_p))
 
-prev_mesh <- 130:350
+prev_mesh <- 250:550
 plot_df <- data.frame(prevalence = prev_mesh, log_prob = dnbinom(prev_mesh, size = curr_nb$neg_binom_r, prob = curr_nb$neg_binom_p, log = TRUE))
 
 prev_figure <- ggplot(plot_df, aes(x = prevalence, y = log_prob)) +