+++ src/BDSCOD/Types.hs
@@ -2,28 +2,32 @@
 
 module BDSCOD.Types where
 
-import Control.DeepSeq
-import Data.Aeson
-import Epidemic.Types
-import GHC.Generics (Generic)
+-- import Control.DeepSeq
+-- import Data.Aeson
+-- import Epidemic.Types
+-- import GHC.Generics (Generic)
+
+
+-- instance NFData Event
+
+-- instance ToJSON Event
+
+-- instance FromJSON Event
+
+import Epidemic.Types.Parameter
 
 type NumLineages = Double
 
-data Event
-  = Birth
-  | Sample
-  | Occurrence
-  | Catastrophe NumLineages
-  | Disaster NumLineages
-  deriving (Show, Eq, Generic)
+data ObservedEvent
+  = OBirth
+  | OSample
+  | OOccurrence
+  | OCatastrophe NumLineages
+  | ODisaster NumLineages
+  deriving (Show, Eq)
+>>>>>>> master
 
-instance NFData Event
-
-instance ToJSON Event
-
-instance FromJSON Event
-
-type Observation = (Time, Event)
+type Observation = (Time, ObservedEvent)
 
 data NegativeBinomial
   = Zero